ZOD_Base/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==3.2.18
django-daterangefilter==1.0.0
django-db-comments==0.4.1
django-environ==0.10.0
django-import-export==3.1.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
et-xmlfile==1.1.0
filelock==3.10.7
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsons==1.6.3
MarkupPy==1.14
MarkupSafe==2.1.2
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
platformdirs==3.2.0
psycopg2==2.9.6
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
sqlparse==0.2.4
tablib==3.4.0
typish==1.9.3
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
Werkzeug==2.2.3
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
